audio-to-emotion-nlp/pyproject.toml
[tool.poetry]
name = "nlp-pipeline"
version = "0.2.1"
description = "Cross-platform NLP pipeline"
authors = [
  "Monika Stangenberg <231648@buas.nl>",
  "Kamil Lega <232753@buas.nl>",
  "Celine Wu <231265@buas.nl>",
  "Deuza Varela <235065@buas.nl>",
  "Sasha Stacie <230306@buas.nl>"
]
readme = "README.md"
package-mode = false # Set to false for development, needs __init__.py files configured for package mode
#packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "3.10.*"
sentencepiece = "^0.2.0"
python-multipart = "^0.0.9"
sacremoses = "^0.1.1"  # Warning messages
fastapi = "^0.115.0"     # Docker
uvicorn = "^0.27.1"      # Docker
python-dotenv = "^1.0.1" # Docker
yt-dlp = "^2024.4.9"     # SAFER than pinning a version that breaks on Docker

# Core and updated dependencies
aiohttp = "^3.11.2"
aiohappyeyeballs = "^2.6.1"
annotated-types = "^0.7.0"
anyio = "^4.2.0"
assemblyai = "^0.37.0"
attrs = "^23.2.0"
blis = "0.7.9"
catalogue = "2.0.10"
certifi = "2025.1.31"
charset-normalizer = "^3.4.1"
click = "^8.1.8"
cloudpathlib = "0.16.0"
cmudict = "^1.0.32"
colorama = "^0.4.6"
comm = "^0.2.2"
confection = "0.0.4"
contourpy = "1.2.0"
cycler = "^0.12.1"
cymem = "2.0.11"
datasets = "^3.5.0"
debugpy = "^1.8.13"
decorator = "^5.2.1"
dill = "^0.3.8"
evaluate = "^0.4.3"
executing = "^2.2.0"
filelock = "^3.18.0"
fonttools = "^4.57.0"
frozenlist = "^1.5.0"
fsspec = "^2024.5.0"
h11 = "^0.14.0"
httpcore = "^1.0.7"
httpx = "^0.28.1"
huggingface-hub = "0.30.1"
idna = "^3.10"
importlib-metadata = "^8.6.1"
importlib-resources = "^6.5.2"
ipykernel = "^6.29.5"
ipython = "8.18.1"
jedi = "^0.19.2"
jinja2 = "^3.1.3"
joblib = "^1.4.2"
jupyter-client = "^8.6.3"
jupyter-core = "^5.7.2"
kiwisolver = "1.3.1"
langcodes = "^3.5.0"
language-data = "^1.3.0"
markdown-it-py = "^3.0.0"
markupsafe = "^2.1.5"
matplotlib = "3.8.4"
matplotlib-inline = "^0.1.7"
mdurl = "^0.1.2"
mpmath = "^1.3.0"
multidict = "^6.3.2"
multiprocess = "^0.70.16"
murmurhash = "1.0.12"
nest-asyncio = "^1.6.0"
networkx = "3.1"
nltk = "^3.9.1"
numpy = "1.26.4"
packaging = "^24.2"
pandas = "2.2.2"
parso = "^0.8.4"
pillow = "^11.1.0"
platformdirs = "^4.3.0"
preshed = "3.0.9"
prompt-toolkit = "^3.0.50"
propcache = "^0.3.1"
psutil = "^5.9.8"
pure-eval = "^0.2.3"
pyarrow = "^15.0.2"
pygments = "^2.19.1"
pyparsing = "^3.2.3"
pyphen = "^0.17.2"
python-dateutil = "^2.9.0.post0"
pytz = "^2025.2"
pyyaml = "^6.0.2"
pyzmq = "^26.0.3"
requests = "2.32.2"
scikit-learn = "1.4.2"
spacy = "3.7.4"
sympy = "^1.12"
textblob = "^0.18.0"
textstat = "0.7.3"
thinc = "^8.1.0"
tqdm = "4.66.3"
transformers = "4.36.2"
vaderSentiment = "^3.3.2"
pip = "^24.0"
pre-commit = "^4.2.0"
azure-ai-ml = "^1.27.1"
azure-identity = "^1.23.0"
azureml-core = "^1.60.0.post1"
librosa = "^0.11.0"
soundfile = "^0.13.1"
typer = "0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.1"
pytest-xdist = "^3.5.0"
tomli = "^2.2.1"
mypy = "^1.15.0"
docformatter = "^1.7.7"
pre-commit = "^4.2.0"
isort = "^6.0.1"
autoflake = "^2.3.1"
black = "^25.1.0"
flake8 = "^7.2.0"
flake8-docstrings = "^1.7.0"
pydocstyle = "^6.3.0"
pandas-stubs = { version = "^2.2.0", markers = "python_version >= '3.10' and python_version < '3.11'" }
types-requests = "^2.32.0.20250515"
types-tqdm = "^4.67.0.20250516"
pytest-cov = "^6.1.1"
poetry2conda = "^0.3.0"

[tool.poetry.group.docs.dependencies]
sphinx = "6.2"
sphinx-autobuild = "^2024.10.3"
sphinx-rtd-theme = "^3.0.2"

[build-system]
requires = ["poetry-core>=1.8.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.poetry2conda]
name = "nlp7-emotion-env"
